move smokescreen metrics into its own package
