+++ cmd/smokescreen.go
@@ -264,7 +264,7 @@ func NewConfiguration(args []string, logger *log.Logger) (*smokescreen.Config, e
 
 		// Setup the connection tracker if there is not yet one in the config
 		if conf.ConnTracker == nil {
-			conf.ConnTracker = conntrack.NewTracker(conf.IdleTimeout, conf.MetricsClient.StatsdClient(), conf.Log, conf.ShuttingDown, nil)
+			conf.ConnTracker = conntrack.NewTracker(conf.IdleTimeout, conf.MetricsClient, conf.Log, conf.ShuttingDown, nil)
 		}
 		configToReturn = conf
 		return nil
