+++ pkg/smokescreen/metrics/mocks.go
@@ -1,125 +1,12 @@
-package smokescreen
+package metrics
 
 import (
 	"fmt"
 	"sort"
 	"strings"
 	"sync"
-	"testing"
 	"time"
-
-	"github.com/stretchr/testify/require"
 )
-
-func TestMetricsTags(t *testing.T) {
-	r := require.New(t)
-
-	t.Run("add custom tags", func(t *testing.T) {
-		metric := "acl.allow"
-		mc := NewNoOpMetricsClient()
-
-		err := mc.AddMetricTags(metric, []string{"globalize"})
-		r.NoError(err)
-
-		tags := mc.GetMetricTags(metric)
-		r.Len(tags, 1)
-		r.Equal(tags[0], "globalize")
-
-		err = mc.AddMetricTags(metric, []string{"ignore"})
-		r.NoError(err)
-
-		tags = mc.GetMetricTags(metric)
-		r.Len(tags, 2)
-	})
-
-	t.Run("add invalid tags", func(t *testing.T) {
-		metric := "acl.does.not.exist"
-		mc := NewNoOpMetricsClient()
-
-		err := mc.AddMetricTags(metric, []string{"globalize"})
-		r.Error(err)
-	})
-}
-
-func TestMetricsClient(t *testing.T) {
-	r := require.New(t)
-
-	// Passing NewMetricsClient a missing statsd address should always fail
-	t.Run("nil statsd addr", func(t *testing.T) {
-		mc, err := NewMetricsClient("", "test_namespace")
-		r.Error(err)
-		r.Nil(mc)
-	})
-
-	// MetricsClient is not thread safe. Adding a tag after smokescreen has started
-	// should always return an error.
-	t.Run("adding metrics after started", func(t *testing.T) {
-		mc := NewNoOpMetricsClient()
-		mc.started.Store(true)
-
-		err := mc.AddMetricTags("acl.allow", []string{"globalize"})
-		r.Error(err)
-	})
-}
-
-func TestMockMetricsClient(t *testing.T) {
-	r := require.New(t)
-
-	t.Run("Incr", func(t *testing.T) {
-		m := NewMockMetricsClient()
-		m.Incr("foobar", 1)
-		c, err := m.GetCount("foobar")
-		r.NoError(err)
-		r.Equal(c, uint64(1))
-	})
-
-	t.Run("Multiple Incr", func(t *testing.T) {
-		m := NewMockMetricsClient()
-		m.Incr("foobar", 1)
-		m.Incr("foobar", 1)
-		m.Incr("foobar", 1)
-		c, err := m.GetCount("foobar")
-		r.NoError(err)
-		r.Equal(c, uint64(3))
-
-		m.Incr("foobar", 123)
-		c, err = m.GetCount("foobar")
-		r.NoError(err)
-		r.Equal(c, uint64(4))
-	})
-
-	t.Run("IncrWithTags", func(t *testing.T) {
-		m := NewMockMetricsClient()
-		tags := []string{"foo", "bar"}
-		m.IncrWithTags("foobar", tags, 1)
-		c, err := m.GetCount("foobar", tags...)
-		r.NoError(err)
-		r.Equal(c, uint64(1))
-		c, err = m.GetCount("foobar")
-		r.NoError(err)
-		r.Equal(c, uint64(1))
-	})
-
-	t.Run("Timing", func(t *testing.T) {
-		m := NewMockMetricsClient()
-		m.Timing("foobar", time.Second, 1)
-		c, err := m.GetCount("foobar")
-		r.NoError(err)
-		r.Equal(c, uint64(1))
-	})
-
-	t.Run("TimingWithTags", func(t *testing.T) {
-		m := NewMockMetricsClient()
-		tags := []string{"foo", "bar"}
-		m.TimingWithTags("foobar", time.Second, 1, tags)
-		c, err := m.GetCount("foobar", tags...)
-		r.NoError(err)
-		r.Equal(c, uint64(1))
-		c, err = m.GetCount("foobar")
-		r.NoError(err)
-		r.Equal(c, uint64(1))
-	})
-}
 
 // MockMetricsClient is a MetricsClient that counts metric updates.
 type MockMetricsClient struct {
